requirements.txt
pick
pywin32
pyautogui 
pygetwindow 
keyboard